refactor: 컴포넌트 및 아이콘 폴더 구조 변경 및 수정
