--- components/common/ui/Carousel/index.ts
@@ -1,3 +1,0 @@
-export * from './Slider';
-export * from './LandingCarousel';
-export * from './ListCarousel';
